requirements.txt
Django==6.0.4
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.32.3
asgiref>=3.9.1
sqlparse==0.5.1
cryptography==42.0.8
django-cors-headers==4.4.0
pytz==2024.1
